Split out each lint into a separate module
Best viewed with `git diff --color-moved`
